rename and move intake sites
