+++ packages/core/src/domain/configuration/transportConfiguration.spec.ts
@@ -1,4 +1,4 @@
-import { INTAKE_SITE_FED_STAGING } from '../../tools/intakeConstants'
+import { INTAKE_SITE_FED_STAGING } from '../intakeSites'
 import type { Payload } from '../../transport'
 import { computeTransportConfiguration, isIntakeUrl } from './transportConfiguration'
